Refactor wormhole integration test
