+++ wormhole-test/wormhole-integration-test/src/test/java/com/zergclan/wormhole/test/integration/engine/PreciseITEngineTest.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.zergclan.wormhole.test.integration.engine;
+
+import com.zergclan.wormhole.test.integration.framework.param.ParameterizedArrayFactory;
+import com.zergclan.wormhole.test.integration.framework.param.WormholeParameterized;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameters;
+
+import java.util.Collection;
+
+/**
+ * Precise IT engine.
+ */
+@RunWith(Parameterized.class)
+public final class PreciseITEngineTest extends BaseITEngine {
+    
+    public PreciseITEngineTest(final WormholeParameterized parameterized) {
+        super(parameterized);
+    }
+    
+    @Parameters(name = "{0}")
+    public static Collection<WormholeParameterized> getParameters() {
+        return ParameterizedArrayFactory.getInstance().getParameterizedArray();
+    }
+    
+    @Test
+    public void assertAction() {
+        getAssertAction().doAssertAction();
+    }
+}
